integrate most types directly into the components unless they are shared between multiple components
