+++ src/components/StyledComponent/adornments.tsx
@@ -3,7 +3,7 @@ import { InputAdornment } from '@mui/material'
 import SearchIcon from '../../components/Icons/Search'
 import ShowHideEyeIcon from '../../components/Icons/ShowHideEye'
 import DownArrowFilledIcon from '../../components/Icons/DownArrowFilled'
-import { AdornmentProps } from '../../types/styledcomponent'
+import { AdornmentProps } from './index'
 
 const StartAdornment: React.FC<AdornmentProps> = props => {
   const { componentvariant } = props
