+++ src/components/StyledComponent/index.tsx
@@ -1,6 +1,11 @@
-import React, { useCallback, useState, useRef, useEffect } from 'react'
+import React, {
+  useCallback,
+  useState,
+  useRef,
+  useEffect,
+  RefObject,
+} from 'react'
 import { Box, InputLabel, OutlinedInput } from '@mui/material'
-import { StyledComponentProps } from '../../types/styledcomponent'
 import { useDropdown } from '../../hooks/styledcomponent/useDropdown'
 import { usePhoneNumber } from '../../hooks/styledcomponent/usePhoneNumber'
 import { usePassword } from '../../hooks/styledcomponent/usePassword'
@@ -13,6 +18,82 @@ import { helperFooterAtom } from '../../atoms/helperfooter'
 import { HelperFooterMessage } from '../../types/validation'
 import { debounce } from 'lodash'
 import labelStyles from '../../styles/StyledComponent/Label'
+import { columnconfig } from '../../components/Grid'
+
+declare module '@mui/material/OutlinedInput' {
+  interface OutlinedInputPropsColorOverrides {
+    [key: string]: true
+  }
+}
+
+declare module '@mui/material/FormControl' {
+  interface FormControlPropsColorOverrides {
+    [key: string]: true
+  }
+}
+
+declare module '@mui/material/InputBase' {
+  interface InputBasePropsColorOverrides {
+    [key: string]: true
+  }
+}
+
+export interface StyledComponentProps {
+  name?: string
+  outlinecolor?: string
+  iconcolor?: string
+  backgroundcolor?: string
+  notched?: boolean
+  combinedfontcolor?: string
+  unshrunkfontcolor?: string
+  shrunkfontcolor?: string
+  endAdornmentMarginRight?: number | string
+  autoComplete?: string
+  componentvariant?:
+    | 'multilinetextfield'
+    | 'dropdown'
+    | 'searchbar'
+    | 'textfield'
+    | 'phonenumber'
+    | 'password'
+    | 'ip-address'
+    | 'email'
+    | 'url'
+    | 'credit-card'
+    | 'number'
+    | 'hostname'
+    | 'domain'
+    | 'time'
+    | 'date'
+  options?: readonly string[]
+  helperfooter?: HelperFooterMessage
+  placeholder?: string
+  minRows?: number
+  formname?: string
+  label?: string
+  shrunklabellocation?: 'onnotch' | 'above' | 'left'
+  value?: string
+  onChange?: (
+    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => void
+  defaultValue?: string
+  inputRef?: RefObject<HTMLInputElement>
+  columnconfig?: columnconfig
+  serverActionValidation?: (
+    formData: FormData
+  ) => Promise<HelperFooterMessage | undefined>
+  focused?: boolean
+}
+
+export interface AdornmentProps {
+  componentvariant: string
+  iconcolor?: string
+  passwordVisible?: boolean
+  // eslint-disable-next-line no-unused-vars
+  togglePasswordVisibility?: (event: React.MouseEvent<HTMLDivElement>) => void
+  // eslint-disable-next-line no-unused-vars
+  marginRight?: number | string
+}
 
 const StyledComponent: React.FC<StyledComponentProps> = props => {
   const {
